refactor/2 attendance command/query/analytics 서비스 분리 적용
